Rename wayland-client.hpp -> wayland-client-core.hpp, add get_interface_name()
Rename is necessary to break dependency cycle between wayland-client.hpp
and wayland-client-protocol.hpp which prohibits adding additional
functionality to wayland-client.hpp that depends on the protocol classes
already being available.
